requirements.txt
python-box==7.0.1
pycocotools==2.0.7
opencv_python==4.7.0.72
Pillow==9.3.0
lightning==2.0.1
segmentation-models-pytorch==0.3.2
albumentations==1.3.1
imagecorruptions==1.1.2
safetensors==0.4.1
torchsummary==1.5.1
tensorboard==2.14.0
einops

